requirements.txt
cycler==0.10.0
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.1
nibabel==2.5.1
numpy==1.17.2
pandas==0.25.1
pickle5==0.0.8
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
xlrd==1.2.0
